requirements.txt
streamlit
Pillow
openai==0.28.1
streamlit-lottie
requests
